ai-cyber-protecting-app/backend/requirements.txt
Flask==2.3.3
openai==1.3.8
python-dotenv==1.0.0
Flask-Cors==4.0.0
requests==2.31.0
